requirements.txt
bcrypt==4.1.2
boto3==1.34.83
botocore==1.34.83
jmespath==1.0.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
